Added new rules and features
The library supports arguments being passed with the validation type

maxLength|5

Arguments should be seperated using the pipe symbol.

All the arguments are passed as an array. The first argument [0] is the name of the rule

As many arguments as required can be passed in this way. It is up to the
rule function to make use of the arguments in the array.

- maxLength
- minLength
- length